refactor: use integer math for animation timestamp

+++ src/libs/gltf/lv_gltf_view/lv_gltf_view_internal.h
@@ -67,7 +67,7 @@ typedef struct {
     bool frame_was_antialiased;
     bool dirty;
     int32_t camera;
-    float animation_speed_ratio;
+    int32_t animation_speed_ratio;
     lv_gltf_antialiasing_mode_t aa_mode;
     lv_gltf_background_mode_t bg_mode;
     float blur_bg;              /** How much to blur the environment background, between 0.0 and 1.0 */
